area_callback/docker-requirements.txt
bosdyn-api>=3.3.0
    # via
    #   bosdyn-client
    #   bosdyn-core
bosdyn-client>=3.3.0
    # via -r requirements.txt
bosdyn-core>=3.3.0
    # via bosdyn-client
certifi>=2022.5.18.1
    # via requests
charset-normalizer>=2.0.12
    # via requests
deprecated>=1.2.13
    # via
    #   bosdyn-client
    #   bosdyn-core
grpcio>=1.46.3
    # via bosdyn-client
idna>=3.3
    # via requests
numpy>=1.19.4
    # via bosdyn-client
protobuf>=3.19.4
    # via bosdyn-api
pyjwt>=2.4.0
    # via bosdyn-client
requests>=2.27.1
    # via bosdyn-client
six>=1.16.0
    # via grpcio
urllib3>=1.26.9
    # via requests
wrapt>=1.14.1
    # via deprecated